Ctrl Advice and exception handler

+++ src/main/java/com/java6/airlineservice/airlineservice/repository/ScheduleRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 public interface ScheduleRepository extends JpaRepository<Schedule,Long> {
 
     List<Schedule> findAllByFromAirportCodeAndToAirportCodeAndDeptTimeBetweenAndRemCapacityGreaterThanEqual(String fromAirportCode, String toAirportCode,
-                                                                              LocalDateTime flightTimeFrom, LocalDateTime flightTimeTo, Long numberOfPeople);
+                                                                              LocalDateTime flightTimeFrom, LocalDateTime flightTimeTo, Integer numberOfPeople);
 
 
 }
